+++ CHANGELOG.md
@@ -48,6 +48,7 @@ v5.1.0
 - Add QCriterion for 2D flow (!1243)
 - Fix to interppointsdatatofld to allow for mpi processing of large files (!1191)
 - Fix the logic of C0Projection:helmsmoothing (!1220)
+- Fix extract module for boundaries with periodic boundary conditions (!1277)
 
 **CardiacEPSolver**
 - Added additional parameter sets to Fenton-Karma model (!1119)